Adding DC Design Week to featured sites
DC Design Week is led and produced by AIGA DC, the local chapter of the professional association for design. AIGA advances design as a professional craft, strategic advancement, and vital cultural force.

The DC chapter was founded in 1984, and is run by a volunteer board of directors. With over 1,230 members, AIGA DC is the fifth largest and one of the oldest chapters in the nation. We strive to cultivate, connect and celebrate the diverse work and people that make up our DC creative community.