Slimming the build
Removing unused partials and code to reduce size of build to less than 50kb.

Ideally we should just run purify/uncss in the project repo so we can keep a more complete build of Tachyons. For now, this will do. I have tried to ensure that we only keep the styles that we need.