Update from tachyons lib (#6)
* Add block elements to box-sizing module

Include `aside`, `nav`, `blockquote`,  `figcaption`, and `figure` block elements in the box-sizing module.

* 4.9.1

* Slim tachyons build

Remove tachyons modules that are not in use

* Regenerated css

* Update package.json

* Slimming the build

Removing unused partials and code to reduce size of build to less than 50kb.

Ideally we should just run purify/uncss in the project repo so we can keep a more complete build of Tachyons. For now, this will do. I have tried to ensure that we only keep the styles that we need.

* Rename output css file